Split Alchemi calculator module into energy, FIRE, and shared files
alchemi_calculator.py held both AlchemiCalculator (energy-only) and
AlchemiFCalculator (FIRE relaxation) plus ~210 lines of helpers shared
by both. Split for readability, mirroring the MACE energy/F file pair:

- _alchemi_common.py     shared internals (model loading, batching,
                         neighbor-list bootstrap, FixAtoms, Langevin MD)
- alchemi_calculator.py  AlchemiCalculator only
- alchemi_f_calculator.py  AlchemiFCalculator + FIRE optimizer registry

Code moved verbatim; no behavior change. __init__ re-exports both names,
so all call sites (package-level imports) are unaffected.
